rename: payment api 패키지명 수정
- 일관성을 위해 controller -> external api로 수정
