+++ loopz-payment-service/src/main/java/kr/co/loopz/payment/apiExternal/TestController.java
@@ -1,4 +1,4 @@
-package kr.co.loopz.payment.controller;
+package kr.co.loopz.payment.apiExternal;
 
 import io.portone.sdk.server.common.Currency;
 import kr.co.loopz.payment.dto.response.PurchasedItem;
