+++ loopz-payment-service/src/main/java/kr/co/loopz/payment/client/ProductClientAtPayment.java
@@ -10,10 +10,10 @@ import org.springframework.web.bind.annotation.RequestHeader;
 import java.util.List;
 
 @FeignClient(
-        name = "product-service-client",
+        name = "product-service-client-at-payment",
         url = "${etc.product-service-url}"
 )
-public interface ProductServiceClient {
+public interface ProductClientAtPayment {
 
     /**
      * 결제된 상품의 재고를 감소시키고, 장바구니에 있다면 상품을 삭제합니다.
